+++ front/report.dynamic.php
@@ -34,6 +34,7 @@
  */
 
 use Glpi\Exception\Http\AccessDeniedHttpException;
+use Glpi\Plugin\Hooks;
 
 if (!isset($_GET['item_type']) || !is_string($_GET['item_type']) || !is_a($_GET['item_type'], CommonGLPI::class, true)) {
     return;
@@ -110,7 +111,7 @@ if (isset($_GET["display_type"])) {
         default:
            // Plugin case
             if ($plug = isPluginItemType($itemtype)) {
-                if (Plugin::doOneHook($plug['plugin'], 'dynamicReport', $_GET)) {
+                if (Plugin::doOneHook($plug['plugin'], Hooks::AUTO_DYNAMIC_REPORT, $_GET)) {
                     return;
                 }
             }